+++ LiteNgine/core/vulkan/Reworked/LtMesh.h
@@ -53,7 +53,7 @@ namespace lte {
 			model = glm::scale(model, -scale);
 			return model;
 		}
-		uint16_t NodeIndex = -1;
+		uint16_t NodeIndex = static_cast<uint16_t>(-1);
 	};
 	struct Vertex
 	{
@@ -236,7 +236,7 @@ namespace lte {
 			model = glm::scale(model, -scale);
 			return model;
 		}
-		uint16_t NodeIndex = -1;
+		uint16_t NodeIndex = static_cast<uint16_t>(-1);
 	};
